fix datatypes

+++ TrainingApp.BL/DTO/KickDTO.cs
@@ -8,8 +8,8 @@ namespace TrainingApp.BLL.DTO
         public long ExcerciseId { get; set; }
         public long SensorId { get; set; }
         public DateTime KickDateTime { get; set; }
-        public decimal? KickPower { get; set; }
-        public decimal? ReactionSpeed { get; set; }
+        public double? KickPower { get; set; }
+        public double? ReactionSpeed { get; set; }
         public int? KickAccuracy { get; set; }
     }
 }
